refactor: #111 - 연산 프로퍼티, 주석 정리(Onboarding,DatePicker, Login, PDF, Policy) - ViewModel 제외

+++ BestWish/Presentation/Scene/MyPage/ProfileUpdate/ViewController/Subview/ProfileUpdateView.swift
@@ -143,7 +143,7 @@ private extension ProfileUpdateView {
         _confirmChangeButton.snp.makeConstraints { 
             $0.bottom.equalTo(safeAreaLayoutGuide).inset(12)
             $0.directionalHorizontalEdges.equalToSuperview().inset(20)
-            $0.height.equalTo(54)
+            $0.height.equalTo(CGFloat(54))
         }
     }
 }
